Attempt at sliding window
Made an attempt at sliding window. Still not getting any data
transferred (resultant file is 0 bytes).
